refactor

+++ database/migrations/2024_11_13_220301_create_document_line_items_table.php
@@ -17,12 +17,12 @@ return new class extends Migration
             $table->morphs('documentable');
             $table->foreignId('offering_id')->nullable()->constrained()->nullOnDelete();
             $table->string('description')->nullable();
-            $table->integer('quantity')->default(1);
-            $table->integer('unit_price')->default(0);
-            $table->integer('subtotal')->storedAs('quantity * unit_price');
-            $table->integer('total')->storedAs('(quantity * unit_price) + tax_total - discount_total');
-            $table->integer('tax_total')->default(0);
-            $table->integer('discount_total')->default(0);
+            $table->decimal('quantity', 10, 2)->default(1);
+            $table->bigInteger('unit_price')->default(0);
+            $table->bigInteger('subtotal')->storedAs('quantity * unit_price');
+            $table->bigInteger('total')->storedAs('(quantity * unit_price) + tax_total - discount_total');
+            $table->bigInteger('tax_total')->default(0);
+            $table->bigInteger('discount_total')->default(0);
             $table->foreignId('created_by')->nullable()->constrained('users')->nullOnDelete();
             $table->foreignId('updated_by')->nullable()->constrained('users')->nullOnDelete();
             $table->timestamps();
